requirements.txt
Django~=5.1.8
django-filter~=24.3
django-bootstrap5~=24.3
django-widget-tweaks~=1.5.0
python-dotenv
dj-database-url
psycopg2-binary
gunicorn
uvicorn
whitenoise
django-softdelete
django-mathfilters

# Dependencies for testing
pytest
pytest-django
pytest-html
pytest-cov
pytest-xdist
pytest-sugar
factory-boy
